Fix build errors in aspnet
